# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cargexo Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Cargexo API documentation"
        }
    ],
    "paths": {
...
        "/all/drivers": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "get all drivers",
                "description": "get all drivers",
                "operationId": "get all drivers",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "filter": {
                                        "type": "array",
                                        "items": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "example": [
                                            [
                                                "endorsements",
                                                "=",
                                                "any"
                                            ],
                                            [
                                                "user_id",
                                                "=",
                                                "2"
                                            ]
                                        ]
                                    },
                                    "columns": {
                                        "type": "array",
                                        "items": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "example": [
                                            "id",
                                            "endorsements",
                                            "dl_exp"
                                        ]
                                    },
                                    "sort": {
                                        "type": "object",
                                        "example": {
                                            "column": "id",
                                            "dir": "desc"
                                        }
                                    },
                                    "skip": {
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "take": {
                                        "type": "string",
                                        "example": "10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/drivers": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "create drivers",
                "description": "create drivers",
                "operationId": "create drivers",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "personal_phone": {
                                        "type": "integer"
                                    },
                                    "business_phone": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "position_id": {
                                        "type": "integer"
                                    },
                                    "company_id": {
                                        "type": "integer"
                                    },
                                    "birth": {
                                        "type": "string"
                                    },
                                    "home_terminal": {
                                        "type": "string"
                                    },
                                    "dl_hash": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "endorsements": {
                                        "type": "string"
                                    },
                                    "hazmat": {
                                        "type": "integer"
                                    },
                                    "tanker": {
                                        "type": "integer"
                                    },
                                    "double_triple": {
                                        "type": "integer"
                                    },
                                    "dl_exp": {
                                        "type": "string"
                                    },
                                    "profile_image": {
                                        "type": "file"
                                    },
                                    "medical_exp": {
                                        "type": "string"
                                    },
                                    "pay_rate": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/drivers/{id}": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "Get drivers By ID",
                "description": "Returns drivers full info",
                "operationId": "get drivers ByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "driverID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/update/drivers/{id}": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "update drivers",
                "description": "update drivers",
                "operationId": "update drivers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "driver ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "personal_phone": {
                                        "type": "integer"
                                    },
                                    "business_phone": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "position_id": {
                                        "type": "integer"
                                    },
                                    "company_id": {
                                        "type": "integer"
                                    },
                                    "birth": {
                                        "type": "string"
                                    },
                                    "home_terminal": {
                                        "type": "string"
                                    },
                                    "dl_hash": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "endorsements": {
                                        "type": "string"
                                    },
                                    "hazmat": {
                                        "type": "boolean"
                                    },
                                    "tanker": {
                                        "type": "boolean"
                                    },
                                    "double_triple": {
                                        "type": "boolean"
                                    },
                                    "dl_exp": {
                                        "type": "string"
                                    },
                                    "profile_image": {
                                        "type": "file"
                                    },
                                    "medical_exp": {
                                        "type": "string"
                                    },
                                    "pay_rate": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/update/drivers/{id}/status": {
            "post": {
                "tags": [
                    "Driver"
                ],
                "summary": "update drivers status ",
                "description": "update drivers status",
                "operationId": "update drivers status",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "driversID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/delete/drivers/{id}": {
            "delete": {
                "tags": [
                    "Driver"
                ],
                "summary": "delete drivers ",
                "description": "delete drivers",
                "operationId": "delete driver",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "driver ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}